Enhance to better handle scanning a directory vs. a specified file
When deluge is used, a specific file is identified and should be used as a priority. But if no deluge arguments are provided, then the specified directory should be scanned instead. Right now the directory scanning requires a complex set of specific arguments to enable- this change should make it so that everything necessary is retrieved from the configs by default.